feat: refactor user authentication link service and introduce types for better clarity
