+++ src/domain/community/user/user.service.ts
@@ -65,10 +65,8 @@ import { AgentInfoCacheService } from '@core/authentication.agent.info/agent.inf
 import { VisualType } from '@common/enums/visual.type';
 import { InstrumentService } from '@src/apm/decorators';
 import { CreateUserSettingsInput } from '../user-settings/dto/user.settings.dto.create';
-import {
-  UserAuthenticationLinkOutcome,
-  UserAuthenticationLinkService,
-} from './user.authentication.link.service';
+import { UserAuthenticationLinkService } from './user.authentication.link.service';
+import { UserAuthenticationLinkOutcome } from './user.authentication.link.types';
 
 @InstrumentService()
 @Injectable()
